requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
slowapi>=0.1.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
